requirements.txt
rq
sqlmodel
sqlalchemy
pydantic
python-dotenv
pydantic_settings
loguru
black
isort
fastapi
requests
apscheduler
uvicorn
psycopg2-binary